+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Support `decimal` and `double` types for tarantool 2.3
+
 ### Fixed
 
 - Removed unnecessary logs